requirements.txt
Pillow==8.4.0
pandas==1.3.4
matplotlib==3.4.3
openpyxl==3.0.9
